bin/channel/requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Google Gemini (REQUIRED for Consultant)
google-generativeai>=0.3.0

# Async HTTP client
aiohttp>=3.9.0

# File watching (optional, for Cursor integration)
watchdog>=3.0.0

# Redis/Celery (optional, for async processing)
# redis>=5.0.0
# celery>=5.3.0
